split headers different, simple cast fixes (#8928)

+++ lib/V8/v8-utils.cpp
@@ -24,7 +24,9 @@
 #include "v8-utils.h"
 
 #ifdef _WIN32
+#include <windef.h>
 #include <conio.h>
+#include <WinSock2.h>
 #include "Basics/win-utils.h"
 #endif
 
@@ -4130,7 +4132,7 @@ static void convertProcessInfoToV8(v8::FunctionCallbackInfo<v8::Value> const& ar
 
   v8::Handle<v8::Array> arguments =
       v8::Array::New(isolate, static_cast<int>(external_process._numberArguments));
-  for (size_t i = 0; i < external_process._numberArguments; i++) {
+  for (uint32_t i = 0; i < external_process._numberArguments; i++) {
     arguments->Set(i, TRI_V8_ASCII_STRING(isolate, external_process._arguments[i]));
   }
   result->Set(TRI_V8_ASCII_STRING(isolate, "arguments"), arguments);
@@ -5308,7 +5310,7 @@ static void JS_ErrorNumberToHttpCode(v8::FunctionCallbackInfo<v8::Value> const& 
   }
 
   auto num = TRI_ObjectToInt64(isolate, args[0]);
-  auto code = arangodb::GeneralResponse::responseCode(num);
+  auto code = arangodb::GeneralResponse::responseCode(static_cast<int>(num));
 
   using Type = typename std::underlying_type<arangodb::rest::ResponseCode>::type;
   TRI_V8_RETURN_INTEGER(static_cast<Type>(code));
